Update on forgot password

+++ controllers/courses.js
@@ -66,8 +66,8 @@ exports.addCourse = asyncHandler(async (req, res, next ) => {
         if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
           return next(
             new ErrorResponse(
-              `User ${req.user.id} is not authorized to delete this bootcamp`,
-              401
+              `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+              401    
             )
           );
         }
@@ -85,22 +85,34 @@ exports.addCourse = asyncHandler(async (req, res, next ) => {
 // // @access  Private
 
 exports.updateCourse = asyncHandler(async (req, res, next ) => {
+  let course = await Course.findById(req.params.id);
 
-      let course = await Course.findById(req.params.id);
+  if (!course) {
+    return next(
+      new ErrorResponse(`No course with the id of ${req.params.id}`),
+      404
+    );
+  }
 
-      if(!course){
-        return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404);
-      }
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
 
-      course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-      });
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
 
-      res.status(200).json({
-      success: true,
-      data: course
-    });
+  res.status(200).json({
+    success: true,
+    data: course,
+  });
 });
 
 
@@ -110,20 +122,29 @@ exports.updateCourse = asyncHandler(async (req, res, next ) => {
 // // @access  Private
 
 exports.deleteCourse = asyncHandler(async (req, res, next ) => {
+  const course = await Course.findById(req.params.id);
 
-      const course = await Course.findById(req.params.id);
+  if (!course) {
+    return next(
+      new ErrorResponse(`No course with the id of ${req.params.id}`),
+      404
+    );
+  }
 
-      if(!course){
-        return next(
-          new ErrorResponse(`No course with the id of ${req.params.id}`),
-           404
-          );
-      }
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    );
+  }
+    
+  await course.deleteOne();
 
-    await course.deleteOne();
-
-      res.status(200).json({
-      success: true,
-      data: {}
-    });
+  res.status(200).json({
+    success: true,
+    data: {},
+  });
 });
